% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{alignCDSSetWithAlignedAAsAsGuide}
\alias{alignCDSSetWithAlignedAAsAsGuide}
\title{Uses an amino acid multiple sequence alignment to guide the alignment of
codon sequence 'cds'.}
\usage{
alignCDSSetWithAlignedAAsAsGuide(unaligned.cds.set, aligned.aa.set)
}
\arguments{
\item{'unaligned.cds.set'}{an instance of Biostrings::DNAStringSet
representing the unaligned codon sequences}

\item{'aligned.aa.set'}{an instance of Biostrings::AAMultipleAlignment
representing the aligned amino acid sequences used as guide}
}
\value{
An instance of Biostrings::DNAMultipleAlignment representing the
aligned codon sequences.
}
\description{
Uses an amino acid multiple sequence alignment to guide the alignment of
codon sequence 'cds'.
}

