% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{alignCodingSequencesPipeline}
\alias{alignCodingSequencesPipeline}
\title{Runs the pipeline to align a set of coding sequences: First translates them,
then validates them for premature stop codons, subsequently generates a
multiple sequence alignment (MSA) of amino acid (AA) sequences, then uses
this AA MSA as guide and aligns the coding sequences in the final step.}
\usage{
alignCodingSequencesPipeline(cds, work.dir, gene.group.name)
}
\arguments{
\item{cds}{an instance of Biostrings::DNAStringSet representing the coding
sequences that need to be aligned}

\item{work.dir}{the working directory to use and in which to save the
relevant files}

\item{gene.group.name}{a string being used to name the output files written
into work.dir. Could be something like 'fam1234'.}
}
\value{
The ALIGNED and validated coding sequences as an instance of
Biostrings::DNAStringSet, or nothing if validation discards the rest of
'cds'.
}
\description{
Runs the pipeline to align a set of coding sequences: First translates them,
then validates them for premature stop codons, subsequently generates a
multiple sequence alignment (MSA) of amino acid (AA) sequences, then uses
this AA MSA as guide and aligns the coding sequences in the final step.
}

