% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{computeKsPipeline}
\alias{computeKsPipeline}
\title{Runs the pipeline to compute the Ks values of gene pair 'x' with coding
sequences held in 'cds'. Temporary files are stored in 't.d'.}
\usage{
computeKsPipeline(x, cds, t.d = tempdir())
}
\arguments{
\item{x}{character vector of gene IDs}

\item{cds}{an instance of Biostrings::XStringSet holding the unaligned
coding sequences of the genes in 'x'}

\item{t.d}{the directory to store the files in, default is tempdir()}
}
\value{
A numeric being the result of invoking 'KaKs_Calculator' on the
aligned coding sequences, or NA if an error occurres.
}
\description{
Runs the pipeline to compute the Ks values of gene pair 'x' with coding
sequences held in 'cds'. Temporary files are stored in 't.d'.
}

