% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{familyDist}
\alias{familyDist}
\title{Generates a distance matrix for genes in family 'fam.name' with distance
values (Ks) obtained from 'ks.tbl'.}
\usage{
familyDist(fam.name, fam.tbl = chi.paralogous.fams,
  ks.tbl = chi.paranome.ks, na.as.ks = 1e+06)
}
\arguments{
\item{fam.name}{The name of the gene family to generate a distance matrix
for}

\item{fam.tbl}{An instance of base::data.frame with two columns. Columns one
must hold the family names and column two the gene members (IDs). Default is
'chi.paralogous.fams'.}

\item{ks.tbl}{An instance of base::data.frame with three columns. Column one
and two hold gene identifier, and column three the measured distance (Ks)
value. Default is 'chi.paranome.ks'.}

\item{na.as.ks}{The value to use, if for a given gene pair no distance value
can be found in 'ks.tbl'. Default is 1e6}
}
\value{
An instance of base::matrix(numeric) holding the pairwise distances
(Ks) of the genes in family 'fam.name'. Obviously the matrix' diagonal will
be zero.
}
\description{
Generates a distance matrix for genes in family 'fam.name' with distance
values (Ks) obtained from 'ks.tbl'.
}

