% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{filterMSA}
\alias{filterMSA}
\title{Filters a multiple coding sequence alignment discarding positions where any
sequence has a gap or a 'N' character.}
\usage{
filterMSA(cds.msa, chars.2.discard.regex = "[nN-]",
  seq.set.type = DNAStringSet)
}
\arguments{
\item{cds.msa}{an instance of Biostrings::XStringSet representing the
multiple sequence alignment}

\item{chars.2.discard.regex}{a string representing the regular expression
defining the character class of the positions to be discarded. Default is
'[nN-]'}

\item{seq.set.type}{one of either Biostrings::DNAStringSet or
Biostrings::AAStringSet. Default is Biostrings::DNAStringSet.}
}
\value{
The filtered XStringSet with the possibly reduced sequences.
}
\description{
Filters a multiple coding sequence alignment discarding positions where any
sequence has a gap or a 'N' character.
}

