% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{fun.bd}
\alias{fun.bd}
\title{Body of inline Rcpp function 'familyDistCpp'}
\format{An object of class \code{character} of length 1.}
\usage{
fun.bd
}
\description{
Body of inline Rcpp function 'familyDistCpp'
}
\keyword{datasets}

