% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getDescendantTipsOrSelf}
\alias{getDescendantTipsOrSelf}
\title{Extracts the tip labels of the tips found in the subtree spanned by 'tr.nd'
in the tree 'phylo.tr'. If the argument node is a tip its label is returned.}
\usage{
getDescendantTipsOrSelf(phylo.tr, tr.nd)
}
\arguments{
\item{phylo.tr}{an instance of ape::phylo representing the tree}

\item{tr.nd}{an integer identifying the node of interest}
}
\value{
A character vector of tip labels
}
\description{
Extracts the tip labels of the tips found in the subtree spanned by 'tr.nd'
in the tree 'phylo.tr'. If the argument node is a tip its label is returned.
}

