% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{isTreeTip}
\alias{isTreeTip}
\title{Checks wether node 'tr.nd' is a tip or not in tree 'phylo.tr'.}
\usage{
isTreeTip(phylo.tr, tr.nd)
}
\arguments{
\item{phylo.tr}{an instance of ape::phylo representing the tree in which to
look up the node.}

\item{tr.nd}{an integer identifying the node for which to determine wether
it is as tip or not.}
}
\value{
TRUE if and only if 'tr.nd' is a tip of tree 'phylo.tr', FALSE
otherwise.
}
\description{
Checks wether node 'tr.nd' is a tip or not in tree 'phylo.tr'.
}

