% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{sanitizeNames}
\alias{sanitizeNames}
\title{Many programs in Bioinformatics cannot handle gene accessions with special
characters. That is why we rename the genes of an Biostrings::XStringSet
with 'PROT1', 'PROT2'...}
\usage{
sanitizeNames(xstring.set)
}
\arguments{
\item{'xstring.set'}{an instance of Biostrings::XStringSet}
}
\value{
An instance of data.frame with two columns the 'original' and the
'sanitized' names.
}
\description{
Many programs in Bioinformatics cannot handle gene accessions with special
characters. That is why we rename the genes of an Biostrings::XStringSet
with 'PROT1', 'PROT2'...
}

