% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{testWeightedDistsForTree}
\alias{testWeightedDistsForTree}
\title{Test function using RUnit to verify the mean distance values for inner tree
nodes.}
\usage{
testWeightedDistsForTree()
}
\arguments{
\item{NONE}{}
}
\value{
TRUE if and only if no validation has failed.
}
\description{
Test function using RUnit to verify the mean distance values for inner tree
nodes.
}

