% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{validateAAStringSet}
\alias{validateAAStringSet}
\title{Validates an AAStringSet}
\usage{
validateAAStringSet(aa.set)
}
\arguments{
\item{'aa.seq'}{an instance of Biostrings::AAStringSet}
}
\value{
A logical vector indicating the indices of valid AA-Seqs in the argument set.
}
\description{
Validates an AAStringSet
}

