% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{weightedDistsForFamily}
\alias{weightedDistsForFamily}
\title{Pipeline to compute the weighted distances (Ks) for a family of genes.
Weighting is done based on each bipartition of subtrees at a given internal
node of the generated family tree, such that for each node the mean of the
distances is computed for all gene pairs, with one member from each subtree.}
\usage{
weightedDistsForFamily(fam.name, fam.tbl = chi.paralogous.fams,
  dist.tbl = chi.paranome.ks, na.dist.as.num = 1e+06)
}
\arguments{
\item{fam.name}{the name of the family as it appears in 'fam.tbl'}

\item{fam.tbl}{a two column instance of base::data.frame where the first
column holds the family name and the second the identifier of gene members.
Default is 'chi.paralogous.fams'}

\item{dist.tbl}{an instance of base::data.frame with four columns. Column
one and two hold the gene identifier, column three the distance 'Ks' value,
and column four the concatonated and alphabetically sorted gene IDs joined
by '_'. Default is 'chi.paranome.ks'}

\item{na.dist.as.num}{A large value to be used for gene pairs where, because
of no significant similarity, 'dist.tbl' has no entry.}
}
\value{
A list with two named entries: 'cluster' an instance of ape::phylo
representing the families binary distance tree, and 'weighted.distances' a
named numeric vector with names 'node_i' and values the weighted distance
computed for the corresponding node.
}
\description{
Pipeline to compute the weighted distances (Ks) for a family of genes.
Weighting is done based on each bipartition of subtrees at a given internal
node of the generated family tree, such that for each node the mean of the
distances is computed for all gene pairs, with one member from each subtree.
}

