% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{weightedDistsForNode}
\alias{weightedDistsForNode}
\title{Identifies all given distances between relevant tips. Relevant means that
all gene pairs are considered between the left descending subtree and the
right descending one. This means (1) 'fam.tr' is expected to be a binary
tree in which each node has only two descending branches, and (2) the
weighted distance for a given node is computed from all pairs generated
between left and right descending subtrees. For more details see
Maere, Steven, Stefanie De Bodt, Jeroen Raes, Tineke Casneuf, Marc Van
Montagu, Martin Kuiper, and Yves Van de Peer. “Modeling Gene and Genome
Duplications in Eukaryotes.” Proceedings of the National Academy of Sciences
of the United States of America 102, no. 15 (April 12, 2005): 5454–59.
doi:10.1073/pnas.0501102102.}
\usage{
weightedDistsForNode(fam.tr, tr.nd, ks.tbl = chi.paranome.ks,
  weight.func = getOption("ksParanomeR.dist.weight.func", stats::mean))
}
\arguments{
\item{fam.tr}{an instance of ape::phylo representing the tree of the
currently investigated gene family. The tree MUST be a binary tree in which
each inner node has exactly two descending branches.}

\item{tr.nd}{The inner tree node for which to compute the weighted distances
(Ks).}

\item{ks.tbl}{an instance of base::data.frame with three columns, in which
the first two hold gene identifier and the third holds the measured distance
(Ks) values. Default is 'chi.paranome.ks'.}

\item{weight.func}{The function used to compute the weighted distance at the
family tree's node 'tr.nd'. Default is stats::mean. Other functions can be
set as an option, use
'options("ksParanomeR.dist.weight.func"=yourFunction)'.}
}
\value{
A numeric value computed by invoking 'weight.func' on all distances
of "relevant" gene pairs.
}
\description{
Identifies all given distances between relevant tips. Relevant means that
all gene pairs are considered between the left descending subtree and the
right descending one. This means (1) 'fam.tr' is expected to be a binary
tree in which each node has only two descending branches, and (2) the
weighted distance for a given node is computed from all pairs generated
between left and right descending subtrees. For more details see
Maere, Steven, Stefanie De Bodt, Jeroen Raes, Tineke Casneuf, Marc Van
Montagu, Martin Kuiper, and Yves Van de Peer. “Modeling Gene and Genome
Duplications in Eukaryotes.” Proceedings of the National Academy of Sciences
of the United States of America 102, no. 15 (April 12, 2005): 5454–59.
doi:10.1073/pnas.0501102102.
}

