% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{weightedDistsForTree}
\alias{weightedDistsForTree}
\title{Generates a named numeric vector in which for each inner node of 'fam.tr'
the weighted distance is held.}
\usage{
weightedDistsForTree(fam.tr, ks.tbl = chi.paranome.ks)
}
\arguments{
\item{fam.tr}{an instance of ape::phylo representing the families tree}

\item{ks.tbl}{an instance of base::data.frame with three columns, in which
the first two hold gene identifier and the third holds the measured distance
(Ks) values. Default is 'chi.paranome.ks'.}
}
\value{
A named numeric vector in which names are 'node_X' and values are
each inner nodes weighted distances.
}
\description{
Generates a named numeric vector in which for each inner node of 'fam.tr'
the weighted distance is held.
}

