% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{computeKsPipeline}
\alias{computeKsPipeline}
\title{Runs the pipeline to compute the Ks values of gene pair 'x' with coding
sequences held in 'cds'. Temporary files are stored in 't.d'.}
\usage{
computeKsPipeline(x, cds = chi.cds, aas = chi.aas, t.d = tempdir(),
  codeml.call = getOption("paranomeKsR.codeml.call", "codeml"))
}
\arguments{
\item{x}{character vector of gene IDs}

\item{cds}{an instance of Biostrings::DNAStringSet holding the unaligned
coding sequences of the genes in 'x'. Default is 'paranomeKsR::chi.cds'.}

\item{aas}{an instance of Biostrings::AAStringSet holding the amino acid
sequences for the genes in 'cds'. You can use Biostrings::translate or the
program MACSE to generate them from the coding sequences. Default is
'paranomeKsR::chi.aas'.}

\item{t.d}{the directory to store the files in, default is tempdir()}

\item{codeml.call}{the call passed to system(...) in order to start codeml.
Default is 'codeml', use option 'paranomeKsR.codeml.call' to set another
default.}
}
\value{
A numeric being the result of invoking PAML's codeml on the aligned
coding sequences, or NA if an error occurres.
}
\description{
Runs the pipeline to compute the Ks values of gene pair 'x' with coding
sequences held in 'cds'. Temporary files are stored in 't.d'.
}
\references{
Yang, Z. and Nielsen, R. (2000) Mol. Biol. Evol., 17, 32-43.
}

